+++ src/TrelloDotNet/Model/Options/GetListOptions/GetListOptions.cs
@@ -1,44 +1,28 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Linq;
 using TrelloDotNet.Control;
 
-namespace TrelloDotNet.Model.Options.GetBoardOptions
+namespace TrelloDotNet.Model.Options.GetListOptions
 {
     /// <summary>
-    /// Options on how and what should be included on the Boards (Example only a few fields to increase performance or more nested data to avoid more API calls)
+    /// Options on how and what should be included on the cards (Example only a few fields to increase performance or more nested data to avoid more API calls)
     /// </summary>
-    public class GetBoardOptions
+    public class GetListOptions
     {
         /// <summary>
-        /// Provide one or more action types there (TrelloDotNet.Model.Webhook.WebhookActionTypes) to get them included with the Board
+        /// All or a comma-separated list of fields. Defaults: badges, checkItemStates, closed, dateLastActivity, desc, descData, due, start, email, idBoard, idChecklists, idLabels, idList, idMembers, idShort, idAttachmentCover, manualCoverAttachment, labels, name, pos, shortUrl, url
         /// </summary>
-        public ActionTypesToInclude ActionsTypes { get; set; }
+        public CardFields CardFields { get; set; }
 
         /// <summary>
-        /// Filter of various Boards (Default: All)
+        /// Whether to return Board object the List is on (Default: False)
         /// </summary>
-        public GetBoardOptionsTypesOfBoardsToInclude TypesOfBoardsToInclude { get; set; } = GetBoardOptionsTypesOfBoardsToInclude.All;
+        public bool IncludeBoard { get; set; }
 
         /// <summary>
-        /// Whether to return Cards on the board (Default: None) [NOTE: This option only works when retrieving a single board. On methods 'GetBoardsForMemberAsync', 'GetBoardsCurrentTokenCanAccessAsync' and 'GetBoardsInOrganization' it have no effect]
+        /// Whether to return Card on the list is on (Default: None)
         /// </summary>
-        public GetBoardOptionsIncludeCards IncludeCards { get; set; }
-
-        /// <summary>
-        /// Whether to return Lists on the board (Default: None)
-        /// </summary>
-        public GetBoardOptionsIncludeLists IncludeLists { get; set; }
-
-        /// <summary>
-        /// Whether to return the Organization of the board (Default: False)
-        /// </summary>
-        public bool IncludeOrganization { get; set; }
-
-        /// <summary>
-        /// Whether to return Labels on the board (Default: False)
-        /// </summary>
-        public bool IncludeLabels { get; set; }
+        public GetListOptionsIncludeCards IncludeCards { get; set; } = GetListOptionsIncludeCards.None;
 
         /// <summary>
         /// What board-fields to include if IncludeBoard are set to True
@@ -46,29 +30,9 @@ namespace TrelloDotNet.Model.Options.GetBoardOptions
         public BoardFields BoardFields { get; set; }
 
         /// <summary>
-        /// What Organization (Workspace)-fields to include if IncludeBoard are set to True
+        /// What Kind of Lists should be included (All, Closed or Open)]
         /// </summary>
-        public OrganizationFields OrganizationFields { get; set; }
-
-        /// <summary>
-        /// What card-fields to include if IncludeCards are set to other than None. all or a comma-separated list of fields. Defaults: badges, checkItemStates, closed, dateLastActivity, desc, descData, due, start, email, idBoard, idChecklists, idLabels, idList, idMembers, idShort, idAttachmentCover, manualCoverAttachment, labels, name, pos, shortUrl, url
-        /// </summary>
-        public CardFields CardFields { get; set; }
-
-        /// <summary>
-        /// Whether to return Plugin object of the card (Default: False)
-        /// </summary>
-        public bool IncludePluginData { get; set; }
-
-        /// <summary>
-        /// What types of boards are returned (Default: All)
-        /// </summary> 
-        public GetBoardOptionsFilter Filter { get; set; } = GetBoardOptionsFilter.All;
-
-        /// <summary>
-        /// Additional Parameters not supported out-of-the-box
-        /// </summary>
-        public List<QueryParameter> AdditionalParameters { get; set; } = new List<QueryParameter>();
+        public ListFilter? Filter { get; set; }
 
         /// <summary>
         /// Order the cards returned by this (only used when multiple cards are returned)
@@ -80,43 +44,27 @@ namespace TrelloDotNet.Model.Options.GetBoardOptions
         /// </summary>
         public List<CardsFilterCondition> CardsFilterConditions { get; set; } = new List<CardsFilterCondition>();
 
+        /// <summary>
+        /// Send these Additional Parameters not supported out-of-the-box (should you need to do something to the query-parameters not yet supported by this API)
+        /// </summary>
+        public List<QueryParameter> AdditionalParameters { get; set; } = new List<QueryParameter>();
+
         internal QueryParameter[] GetParameters()
         {
             List<QueryParameter> parameters = new List<QueryParameter>();
+            if (CardFields != null)
+            {
+                parameters.Add(new QueryParameter("fields", string.Join(",", CardFields.Fields)));
+            }
 
             if (BoardFields != null)
             {
-                parameters.Add(new QueryParameter("fields", string.Join(",", BoardFields.Fields)));
+                parameters.Add(new QueryParameter("board_fields", string.Join(",", BoardFields.Fields)));
             }
 
-            if (CardFields != null)
-            {
-                parameters.Add(new QueryParameter("card_fields", string.Join(",", CardFields.Fields)));
-            }
-
-            if (OrganizationFields != null)
-            {
-                parameters.Add(new QueryParameter("organization_fields", string.Join(",", OrganizationFields.Fields)));
-            }
-
-            if (ActionsTypes != null)
-            {
-                parameters.Add(new QueryParameter("actions", string.Join(",", ActionsTypes.ActionTypes)));
-            }
-
-            if (TypesOfBoardsToInclude != GetBoardOptionsTypesOfBoardsToInclude.All)
-            {
-                var selected = Enum.GetValues(TypesOfBoardsToInclude.GetType()).Cast<Enum>().Where(TypesOfBoardsToInclude.HasFlag).ToArray();
-
-                parameters.Add(new QueryParameter("filter", string.Join(",", selected.Select(x => x.GetJsonPropertyName()))));
-            }
-
-            parameters.Add(new QueryParameter("lists", IncludeLists.GetJsonPropertyName()));
-            parameters.Add(new QueryParameter("organization", IncludeOrganization));
-            parameters.Add(new QueryParameter("filter", Filter.GetJsonPropertyName()));
+            parameters.Add(new QueryParameter("board", IncludeBoard));
             parameters.Add(new QueryParameter("cards", IncludeCards.GetJsonPropertyName()));
-            parameters.Add(new QueryParameter("labels", IncludeLabels ? "all" : "none"));
-            parameters.Add(new QueryParameter("pluginData", IncludePluginData));
+
             parameters.AddRange(AdditionalParameters);
 
             return parameters.ToArray();